+++ spec/arduino_installation_spec.rb
@@ -1,12 +1,6 @@
 require "spec_helper"
 
 RSpec.describe ArduinoCI::ArduinoInstallation do
-  context "force_install_location" do
-    it "is resolvable" do
-      expect(ArduinoCI::ArduinoInstallation.force_install_location).not_to be nil
-    end
-  end
-
   context "autolocate" do
     it "doesn't fail" do
       ArduinoCI::ArduinoInstallation.autolocate